added all parameters in repo method fetchCarsByDate

+++ src/main/java/com/example/bilabonnementeksamen/model/Car.java
@@ -7,7 +7,7 @@ public class Car {
   private int car_vehicle_number;
   private int car_chassis_number;
   private CarModel car_model_id;
-  private int car_price_month;
+  private double car_price_month;
   private Subscription_type subscription_type_id;
   private int car_is_reserved;
 
@@ -15,7 +15,7 @@ public class Car {
   public Car(){}
 
   public Car(int car_vehicle_number, int car_chassis_number, CarModel car_model_id,
-             int car_price_month, Subscription_type subscription_type_id, int car_is_reserved) {
+             double car_price_month, Subscription_type subscription_type_id, int car_is_reserved) {
     this.car_vehicle_number = car_vehicle_number;
     this.car_chassis_number = car_chassis_number;
     this.car_model_id = car_model_id;
@@ -48,11 +48,11 @@ public class Car {
     this.car_model_id = car_model_id;
   }
 
-  public int getCar_price_month() {
+  public double getCar_price_month() {
     return car_price_month;
   }
 
-  public void setCar_price_month(int car_price_month) {
+  public void setCar_price_month(double car_price_month) {
     this.car_price_month = car_price_month;
   }
